[skip ci] complete skeleton
